Add guards to switch on or off Wifi depending on state

+++ OTCamera/helpers/rpi.py
@@ -79,23 +79,28 @@ def reboot():
 
 def wifi_switch_on():
     """Turn on Wi-Fi"""
-    if not config.DEBUG_MODE_ON:
-        call("rfkill unblock wlan", shell=True)
-    if config.USE_RELAY:
-        call("sudo systemctl start sshrelay.service", shell=True)
-        log.write("Started SSH relay server connection")
+    if not status.wifi_on:
+        if not config.DEBUG_MODE_ON:
+            call("rfkill unblock wlan", shell=True)
+
+        if config.USE_RELAY:
+            call("sudo systemctl start sshrelay.service", shell=True)
+            log.write("Started SSH relay server connection")
+        status.wifi_on = True
+        log.write("Wi-Fi on")
+
     led.wifi_on()
-    status.wifi_on = True
-    log.write("Wi-Fi on")
 
 
 def wifi_switch_off():
     """Turn off Wi-Fi"""
-    if not config.DEBUG_MODE_ON:
-        call("rfkill block wlan", shell=True)
-    if config.USE_RELAY:
-        call("sudo systemctl stop sshrelay.service", shell=True)
-        log.write("Stopped SSH relay server connection")
+    if status.wifi_on:
+        if not config.DEBUG_MODE_ON:
+            call("rfkill block wlan", shell=True)
+        if config.USE_RELAY:
+            call("sudo systemctl stop sshrelay.service", shell=True)
+            log.write("Stopped SSH relay server connection")
+        status.wifi_on = False
+        log.write("Wi-Fi off")
+
     led.wifi_off()
-    status.wifi_on = False
-    log.write("Wi-Fi off")
